@@ -17,25 +17,30 @@
   "author": "ViewCrunch",
   "license": "MIT",
   "dependencies": {
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@emotion/cache": "^11.6.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.1.1",
+    "@mui/lab": "^5.0.0-alpha.55",
+    "@mui/material": "^5.1.1",
+    "@mui/styles": "^5.1.1",
     "axios": "^0.25.0",
-    "firebase": "^8.8.0",
-    "firebase-admin": "^9.11.0",
-    "js-cookie": "^3.0.0",
-    "next": "^11.1.3",
-    "notistack": "^1.0.10",
-    "prop-types": "^15.7.2",
+    "firebase": "^9.6.4",
+    "firebase-admin": "^10.0.2",
+    "js-cookie": "^3.0.1",
+    "next": "^12.0.8",
+    "notistack": "^2.0.3",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-firebaseui": "^5.0.2",
-    "react-redux": "^7.2.4",
+    "react-firebaseui": "^6.0.0",
+    "react-redux": "^7.2.6",
     "react-reveal": "^1.2.2",
-    "redux": "^4.1.0",
+    "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.36.0",
+    "redux-thunk": "^2.4.1",
+    "sass": "^1.49.0",
     "uuid": "^8.3.2"
   },
   "devDependencies": {}
