@@ -1,4 +1,4 @@
-import Grid from "@material-ui/core/Grid";
+import Grid from "@mui/material/Grid";
 
 import { Footer } from "@component/layout";
 import { NavBar, Intro, Trending, NewsContainer, ViewsContainer, viewsStyles } from ".";
