@@ -1,7 +1,7 @@
 import Link from "next/link";
 
-import Button from "@material-ui/core/Button";
-import Typography from "@material-ui/core/Typography";
+import Button from "@mui/material/Button";
+import Typography from "@mui/material/Typography";
 
 import { styles } from ".";
 import { NavContainer } from "@component/layout";
