@@ -1,8 +1,8 @@
 import Link from "next/link";
 
-import Grid from "@material-ui/core/Grid";
-import Typography from "@material-ui/core/Typography";
-import TimelineIcon from "@material-ui/icons/Timeline";
+import Grid from "@mui/material/Grid";
+import Typography from "@mui/material/Typography";
+import TimelineIcon from "@mui/icons-material/Timeline";
 
 import { styles } from ".";
 
