@@ -10,42 +10,47 @@ const verifyRefresh = async (myRefresh) => {
 
   if (!token) return "invalid user";
 
-  return await firebaseAdmin
-    .auth()
-    .verifyIdToken(token)
-    .then(async (decodedToken) => {
-      const handle = await firebaseAdmin
-        .auth()
-        .getUser(decodedToken?.uid)
-        .then((user) => user.displayName);
+  // const email = await auth.verifyIdToken(token);
+  // console.log(email);
 
-      if (!handle.startsWith("@")) return { myHandle: handle };
-      if (!handle) throw new TypeError("invalid user");
+  console.log("server", "auth", firebaseAdmin);
+  // return await auth
+  //   .verifyIdToken(token)
+  //   .then(async (decodedToken) => {
+  // const handle = await firebaseAdmin
+  //   .auth()
+  //   .getUser(decodedToken?.uid)
+  //   .then((user) => user.displayName);
 
-      const profile = await firebaseAdmin
-        .firestore()
-        .collection("profile")
-        .doc(handle)
-        .get()
-        .then((snapshot) => ({ ...snapshot.data(), id: snapshot.id }))
-        .catch((error) => {
-          throw new TypeError(error);
-        });
+  // if (!handle.startsWith("@")) return { myHandle: handle };
+  // if (!handle) throw new TypeError("invalid user");
 
-      return {
-        myHandle: handle,
-        myID: profile.id,
-        myTheme: profile.stat?.theme,
-        myNotification: profile.notification?.length,
-        myProfilePicture: profile.profilePicture,
-        myCoverPicture: profile.coverPicture,
-        myDisplayName: profile.displayName,
-        myProfession: profile.profession,
-      };
-    })
-    .catch((err) => {
-      throw new TypeError(err);
-    });
+  // const profile = await firebaseAdmin
+  //   .firestore()
+  //   .collection("profile")
+  //   .doc(handle)
+  //   .get()
+  //   .then((snapshot) => ({ ...snapshot.data(), id: snapshot.id }))
+  //   .catch((error) => {
+  //     throw new TypeError(error);
+  //   });
+
+  // return {
+  //   myHandle: handle,
+  //   myID: profile.id,
+  //   myTheme: profile.stat?.theme,
+  //   myNotification: profile.notification?.length,
+  //   myProfilePicture: profile.profilePicture,
+  //   myCoverPicture: profile.coverPicture,
+  //   myDisplayName: profile.displayName,
+  //   myProfession: profile.profession,
+  // };
+  // })
+  // .catch((err) => {
+  //   console.log("server err", { err });
+
+  //   throw new TypeError(err);
+  // });
 };
 
 export default async (req, res) => {
@@ -54,7 +59,7 @@ export default async (req, res) => {
     if (profile === "invalid user") throw new TypeError("Invalid User");
     return res.status(200).json(profile);
   } catch (error) {
-    // console.log(error);
+    console.log(error);
     return res.status(401).json({});
   }
 };
